crypto-oracle/requirements.txt
cryptography
devtools==0.12.2
fastapi==0.109.0
email-validator
glom==22.1.0
httpx==0.26.0
Jinja2==3.0.3
pycryptodome==3.13.0
# pydantic==1.10.0
pydantic==2.5.3
pydantic_settings==2.1.0
pytest==8.0.0
python-jose==3.3.0
python-multipart
PyYAML==6.0.1
uvicorn[standard]